+++ modules/apps/frontend-icons/frontend-icons-admin-web/bnd.bnd
@@ -1,0 +1,4 @@
+Bundle-Name: Liferay Frontend Icons Admin Web
+Bundle-SymbolicName: com.liferay.frontend.icons.admin.web
+Bundle-Version: 1.0.0
+Web-ContextPath: /frontend-icons-admin-web
