+++ modules/apps/frontend-icons/frontend-icons-admin-web/package.json
@@ -1,0 +1,14 @@
+{
+	"dependencies": {
+		"@clayui/form": "3.32.0",
+		"@clayui/icon": "3.32.0",
+		"react": "16.12.0"
+	},
+	"name": "frontend-icons-admin-web",
+	"scripts": {
+		"build": "liferay-npm-scripts build",
+		"checkFormat": "liferay-npm-scripts check",
+		"format": "liferay-npm-scripts fix"
+	},
+	"version": "1.0.0"
+}
