+++ modules/apps/frontend-icons/frontend-icons-admin-web/src/main/resources/content/Language.properties
@@ -1,0 +1,3 @@
+category.icons-admin=Icons Admin
+icons-admin=Icons Admin
+icons-admin-configuration-name=Icons Configuration
