+++ modules/apps/frontend-icons/frontend-icons-clay/.gitignore
@@ -1,1 +1,1 @@
-/src/main/resources/META-INF/resources/images
+/src/main/resources/META-INF/resources/images/*
