--- modules/apps/frontend-icons/frontend-icons-contributor-extender/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Frontend Icons Contributor Extender
-Bundle-SymbolicName: com.liferay.frontend.icons.contributor.extender
-Bundle-Version: 3.0.0
